Update basics/gloo examples to work better with sphinx-gallery
