+++ examples/plotting/README.rst
@@ -1,2 +1,5 @@
 Plotting
 ========
+
+More example scripts are available in the VisPy repository's
+`example scripts directory <https://github.com/vispy/vispy/tree/main/examples>`_.
